Refactor structure of some classes and methods, add comments to methods to make them easier to understand
