+++ Sources/Chroma/Extensions/File.swift
@@ -10,7 +10,6 @@ import Foundation
 import Files
 
 extension File {
-    
     init(named name: String, at folder: Folder) throws {
         guard let file = try? folder.createFileIfNeeded(at: name) else {
             throw ChromaError.fileCreationFailedAtFolder(path: folder.path, fileName: name)
